에디터 툴바 분리
